+++ projects/plugins/jetpack/tests/php/modules/masterbar/test-class-admin-menu.php
@@ -10,6 +10,7 @@ use Automattic\Jetpack\Dashboard_Customizations\Base_Admin_Menu;
 use Automattic\Jetpack\Status;
 
 require_jetpack_file( 'modules/masterbar/admin-menu/class-admin-menu.php' );
+require_jetpack_file( 'modules/masterbar/admin-menu/class-dashboard-switcher.php' );
 require_jetpack_file( 'tests/php/modules/masterbar/data/admin-menu.php' );
 
 /**
@@ -560,37 +561,5 @@ class Test_Admin_Menu extends WP_UnitTestCase {
 				array( '', 'read', 'test-slug', '', Base_Admin_Menu::HIDE_CSS_CLASS ),
 			),
 		);
-	}
-
-	/**
-	 * Check if the dashboard switcher is registered correctly.
-	 */
-	public function test_register_dashboard_switcher() {
-		global $pagenow;
-		$pagenow = 'edit.php?post_type=feedback';
-
-		$output = static::$admin_menu->register_dashboard_switcher( '' );
-		$this->assertNull( $output );
-
-		$screens = array(
-			'edit.php'                             => 'https://wordpress.com/posts/',
-			'edit.php?post_type=page'              => 'https://wordpress.com/pages/',
-			'edit.php?post_type=jetpack-portfolio' => 'https://wordpress.com/types/jetpack-portfolio/',
-			'edit-tags.php?taxonomy=category'      => 'https://wordpress.com/settings/taxonomies/category/',
-		);
-
-		foreach ( $screens as $screen => $mapping ) {
-			$pagenow  = $screen;
-			$output   = static::$admin_menu->register_dashboard_switcher( '' );
-			$expected = sprintf(
-				'<div id="dashboard-switcher"><h5>%s</h5><p class="dashboard-switcher-text">%s</p><a class="button button-primary dashboard-switcher-button" href="%s">%s</a></div>',
-				__( 'Screen features', 'jetpack' ),
-				__( 'Currently you are seeing the classic WP-Admin view of this page. Would you like to see the default WordPress.com view?', 'jetpack' ),
-				$mapping . static::$domain,
-				__( 'Use WordPress.com view', 'jetpack' )
-			);
-
-			$this->assertEquals( $expected, $output );
-		}
 	}
 }
