+++ projects/plugins/jetpack/tests/php/modules/masterbar/test-class-atomic-admin-menu.php
@@ -10,6 +10,7 @@ use Automattic\Jetpack\Status;
 
 require_jetpack_file( 'modules/masterbar/admin-menu/class-atomic-admin-menu.php' );
 require_jetpack_file( 'tests/php/modules/masterbar/data/admin-menu.php' );
+require_jetpack_file( 'modules/masterbar/admin-menu/class-dashboard-switcher.php' );
 
 /**
  * Class Test_Atomic_Admin_Menu.
